chore: move components to components folder
